pyproject.toml
[project]
name = "security-auth"
version = "1.0.0"
description = "Production-grade authentication and authorization"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    # Core
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "authlib>=1.3.0",
    "pyotp>=2.9.0",
    
    # Web
    "fastapi>=0.109.0",
    "httpx>=0.26.0",
    
    # Utils
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
